pyproject.toml
[project]
name = "investing-mcp"
version = "0.1.0"
description = "MCP server for investing and trading tools"
authors = [
    {name = "Wellington Abreu", email = "wellington.abreu@example.com"}
]
dependencies = [
    "anthropic>=0.51.0",
    "aiohttp>=3.8.0",
    "mcp[cli]>=1.12.0",
    "prompt-toolkit>=3.0.51",
    "python-dotenv>=1.1.0",
    "certifi>=2023.0.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]